rename BottomNavigationItem to BottomNavigationButton
