+++ docs/api/BottomNavigation/BottomNavigationButton.md
@@ -1,4 +1,4 @@
-BottomNavigationItem
+BottomNavigationButton
 =======
