+++ src/BottomNavigation/index.js
@@ -2,4 +2,4 @@
 
 export default from './BottomNavigation';
 export BottomNavigation from './BottomNavigation';
-export BottomNavigationItem from './BottomNavigationItem';
+export BottomNavigationButton from './BottomNavigationButton';
